Introduce File Writer Permits
The File Writer Permit arbitrates writing to the File Writer object.
When such a permit is constructed, the File Writer is locked (not
implemented at the moment). At destruction, it flushes the File Writer
buffer to the file if the flushing threshold has been crossed, and
unlocks the file (not implemented).